pyproject.toml
[tool.poetry]
name = "dogfooding-test-agent"
version = "0.1.0"
description = ""
authors = ["Jakub Novak <jakub@e2b.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
e2b = "^0.2.9"
openai = "^0.28.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
